// src/app/lib/types/types.ts
export type IconType =
    | "portfolio"
    | "menu"
    | "search"
    | "settings"
    | "reset"
    | "ai"
    | "job"
    | "facebook"
    | "instagram"
    | "linkedIn"
    | "github"
    | "twitter"
    | "email"
    | "phone"
    | "web"
    | "facebook-round"
    | "instagram-round"
    | "linkedIn-round"
    | "github-round"
    | "twitter-round"
    | "email-round"
    | "phone-round"
    | "web-round"
    | "swatch"
    | "star"
    | "star-solid"
    | "add"
    | "edit"
    | "delete"
    | "save"
    | "cancel"
    | "signup"
    | "signin"
    | "signout"
    | "send"
    | "chat"
    | "arrowDown"
    | "arrowUp"
    | "arrowLeft"
    | "arrowRight"
    | "undo"
    | "redo"
    | "chevronLeft"
    | "chevronRight"
    | "X"
    | "preview"
    | "upload"
    | "creditcard"
    | "devicePhoneMobile"
    | "chartBarSquare"
    | "magnifyingGlass"
    | "cpuChip"
    | "globe"
    | "check";